+++ web/static/react/assets/PracticeProfilePage-C8v2TqBn.js
@@ -1,1 +1,0 @@
-import{t as e}from"./vendor-react-DRzYtaUD.js";import{t}from"./LegalSkillsProfilePage-DnuGJ_x-.js";var n=e();function r(){return(0,n.jsx)(t,{})}export{r as PracticeProfilePage};